+++ maptools/cfg.py
@@ -1,0 +1,3 @@
+KEY = ""
+
+assert KEY != "", "Config `KEY` first."